requirement.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.2
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
django-lock-tokens==0.2.5
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
psycopg2==2.9.1
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
